# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Diruma",
        "description": "",
        "contact": {
            "email": "email@anda.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/testimony": {
            "get": {
                "tags": [
                    "Testimoni"
                ],
                "summary": "Ambil Semua Data Testimoni",
                "operationId": "2db3ce951dca556cb20b1b243f99754a",
                "responses": {
                    "200": {
                        "description": "Sukses ambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak Ada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Testimoni"
                ],
                "summary": "Tambah Testimoni Baru",
                "operationId": "ae81a04d4b8387b19ed0e3a1b4105ac6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Makanannya sangat enak dan higienis!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Sukses"
                                        },
                                        "Pesan": {
                                            "type": "string",
                                            "example": "Testimoni berhasil ditambahkan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (Bukan Admin)"
                    },
                    "400": {
                        "description": "Validasi Gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/testimony/{id}": {
            "get": {
                "tags": [
                    "Testimoni"
                ],
                "summary": "Lihat Detail Satu Testimoni",
                "operationId": "173794b0252b658cc9672f33d60c1b4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Testimoni",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail Ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Sukses"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Testimoni Tidak Ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Testimoni"
                ],
                "summary": "Update Testimoni",
                "operationId": "35bf06dd047582e4718b2ec7408121f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Saputra"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "makanannya enak enak"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil Update"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    },
                    "403": {
                        "description": "Dilarang"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Testimoni"
                ],
                "summary": "Hapus Testimoni",
                "operationId": "7692fb9ffd209d66525950b337012efd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil Dihapus"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login dengan email dan password untuk dapatkan token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}